Moving errors/errors.go -> annotations/errors.go
